Mouse fix for high-DPI non-macOS systems (e.g. Wayland)
The "DPI hack" for Apple Retina systems is actually necessary for all
high-DPI systems. Enable it unconditionally — it's harmless on
non-high-DPI systems — and clean it up a little. In particular:
- Use SDL_GetRendererOutputSize() instead of SDL_GL_GetDrawableSize(),
  which is not guaranteed to work on non-OpenGL-based renderers.
- Move the display-mode related stuff behind a video.useDesktopMouseCoords
  check: ideally we should be using SDL_GetWindowSize() instead.
- Update the comments to describe the DPI calculation more fully.

Note that, in the soon-to-be-released SDL 2.0.18, the helper function
SDL_RenderWindowToLogical() could probably replace a lot of this.
(Similarly, the use of SDL_SetWindowMouseRect() will make more sense
instead of warping for the software cursor case), but those should wait
until after 2.0.18 is fully released and available.